requirements.txt
fastapi
uvicorn
anthropic
lightgbm
shap
pandas
numpy
scipy
scikit-learn
pydantic
fastmcp
pytest
python-multipart
python-dotenv
google-genai
openai
huggingface_hub
